+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * Fixed: Find path when edge labels are deflated
 * Fixed: Accumulo stream property value in table data length of reference
 * Fixed: Query will now return hidden elements/vertices/edges if FetchHint.INCLUDE_HIDDEN is passed
-* Fixed: Query.hasAuthorization will now match elements whose only use of an authorization is a hidden field
+* Fixed: Query.hasAuthorization will now match elements whose only use of an authorization is that the element is hidden
+* Fixed: Query.hasAuthorization will now match elements whose only use of an authorization is a hidden property
 * Added: Query methods elementIds/vertexIds/edgeIds are overloaded to accept IdFetchHint, which makes it possible to include ids for hidden elements
 * Added: Graph.getExtendedDataInRange in order to bulk load ranges of extended data rows
 * Added: SearchIndex.addExtendedData to allow for reindexing extended data